requirements.txt
accelerate
anthropic
chromadb<0.4.0
docx2pdf
fire
google-generativeai
gymnasium
icalendar
ipykernel
ipython<9.0
ipywidgets
jsonargparse
mysql-connector-python
numpy<2.0
openai
openpyxl
pandas
Pillow
PyMuPDF
PyPDF2
pdf2docx
peft
prettyprinter
pytesseract
python_docx==1.1.0
python-dotenv
randomname
ray
rich
rpyc
scikit_learn
sentence_transformers
tiktoken
transformers
wandb